+++ spice_ev/strategies/flex_window.py
@@ -340,8 +340,7 @@ class FlexWindow(Strategy):
             cs = self.world_state.charging_stations[cs_id]
             sim_vehicle = deepcopy(vehicle)
             cur_time = self.current_time - self.interval
-            max_discharge_power = (sim_vehicle.battery.loading_curve.max_power
-                                   * sim_vehicle.vehicle_type.v2g_power_factor)
+            max_discharge_power = sim_vehicle.battery.unloading_curve.max_power
 
             # check if vehicles can be loaded until desired_soc in connected timesteps
             old_soc = vehicle.battery.soc
@@ -359,7 +358,7 @@ class FlexWindow(Strategy):
 
             # check if vehicle ends up with desired soc, adjust min_soc accordingly
             if not cur_window and window_change >= 1:
-                min_soc = self.DISCHARGE_LIMIT
+                min_soc = vehicle.vehicle_type.discharge_limit
                 max_soc = 1
                 while max_soc - min_soc > self.EPS:
                     discharge_limit = (max_soc + min_soc) / 2
@@ -705,8 +704,7 @@ class FlexWindow(Strategy):
             sim_vehicle = deepcopy(vehicle)
             cs_id = sim_vehicle.connected_charging_station
             cs = self.world_state.charging_stations[cs_id]
-            max_discharge_power = (sim_vehicle.battery.loading_curve.max_power
-                                   * sim_vehicle.vehicle_type.v2g_power_factor)
+            max_discharge_power = sim_vehicle.battery.unloading_curve.max_power
 
             # check if vehicles can be loaded until desired_soc in connected timesteps
             old_soc = vehicle.battery.soc
@@ -725,7 +723,7 @@ class FlexWindow(Strategy):
 
             # check if vehicle ends up with desired soc, adjust min_soc accordingly
             if not cur_window and window_change >= 1:
-                min_soc = self.DISCHARGE_LIMIT
+                min_soc = sim_vehicle.vehicle_type.discharge_limit
                 max_soc = 1
                 while max_soc - min_soc > self.EPS:
                     discharge_limit = (max_soc + min_soc) / 2