+++ spice_ev/strategy.py
@@ -1,4 +1,6 @@
 from copy import deepcopy
+
+from datetime import timedelta
 from importlib import import_module
 from warnings import warn
 
@@ -31,13 +33,13 @@ class Strategy():
         self.world_state = deepcopy(components)
         self.world_state.future_events = []
         self.interval = kwargs.get('interval')  # required
+        self.ts_per_hour = self.interval / timedelta(minutes=60)
         self.current_time = start_time - self.interval
         # relative allowed difference between battery SoC and desired SoC when leaving
         self.margin = 0.1
         self.PRICE_THRESHOLD = 0
         self.ALLOW_NEGATIVE_SOC = False
         self.RESET_NEGATIVE_SOC = False
-        self.V2G_POWER_FACTOR = 0.5
         # check if strategy uses grid signals & enable/disable plotting of schedule or window
         self.uses_schedule = False
         self.uses_window = False
@@ -115,7 +117,10 @@ class Strategy():
                     # connector max power not set
                     connector.cur_max_power = ev.max_power
             elif type(ev) is events.VehicleEvent:
-                vehicle = self.world_state.vehicles[ev.vehicle_id]
+                vehicle = self.world_state.vehicles.get(ev.vehicle_id)
+                if vehicle is None:
+                    # skip events without vehicle
+                    continue
                 # update vehicle attributes
                 for k, v in ev.update.items():
                     setattr(vehicle, k, v)
@@ -202,15 +207,15 @@ class Strategy():
                 avg_power = vehicle.battery.load(self.interval, max_power=power)['avg_power']
                 commands[cs_id] = gc.add_load(cs_id, avg_power)
                 cs.current_power += avg_power
-            elif (vehicle.get_delta_soc() < -self.EPS
+            elif (gc_surplus < -self.EPS
+                    and vehicle.get_delta_soc() < -self.EPS
                     and vehicle.vehicle_type.v2g
-                    and cs.current_power < self.EPS
+                    and gc.current_loads.get(cs_id, 0) < self.EPS
                     and not gc_cheap[cs.parent]):
-                # GC draws power, surplus in vehicle and V2G capable: support GC
-                discharge_power = min(
-                    -gc_surplus,
-                    vehicle.battery.loading_curve.max_power * vehicle.vehicle_type.v2g_power_factor)
-                target_soc = max(vehicle.desired_soc, self.DISCHARGE_LIMIT)
+                # GC draws power, surplus in vehicle,
+                # not currently charging and V2G capable: support GC
+                discharge_power = min(-gc_surplus, vehicle.battery.unloading_curve.max_power)
+                target_soc = max(vehicle.desired_soc, vehicle.vehicle_type.discharge_limit)
                 avg_power = vehicle.battery.unload(
                     self.interval, max_power=discharge_power, target_soc=target_soc)['avg_power']
                 commands[cs_id] = gc.add_load(cs_id, -avg_power)