+++ CHANGELOG.md
@@ -17,9 +17,12 @@ Template:
 - [(#)]()
 ```
 
-## [1.0.2] - Unreleased - 2023-08-21
+## [1.0.2] - Unreleased - 2023-10-17
 
 ### Changed
+- distributed strategy: charge vehicles with any strategy at depot and opportunity stations
+- distributed strategy: battery behavior adapted
+- new strategies: peak_load_window
 - plotting time windows with shaded backgrounds
 
 ## [1.0.1] - Minor Fixes - 2023-08-04